pyproject.toml
[tool.poetry]
name = "ordo"
version = "0.1.0"
description = ""
authors = ["Antonio Alonso"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
beautifulsoup4 = "^4.12.3"
tinydb = "^4.8.2"
prettytable = "^3.11.0"
dateutils = "^0.6.12"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
